LayoutTests/platform/wpe/media/video-unmuted-after-play-holds-sleep-assertion-expected.txt


RUN(video.src = findMediaFile("video", "content/test"))
EVENT(canplaythrough)
RUN(video.play())
EVENT(playing)
EXPECTED (internals.elementIsBlockingDisplaySleep(video) == 'false'), OBSERVED 'true' FAIL
RUN(video.muted = false)
EVENT(volumechange)
EXPECTED (internals.elementIsBlockingDisplaySleep(video) == 'true') OK
END OF TEST
